unfollow

+++ src/main/java/services/FriendService.java
@@ -3,6 +3,7 @@ package services;
 import static com.googlecode.objectify.ObjectifyService.ofy;
 
 import java.util.Date;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 import java.util.logging.Logger;
@@ -18,11 +19,13 @@ import entities.NextCursor;
 import entities.TwitterFriend;
 import twitter4j.PagableResponseList;
 import twitter4j.RateLimitStatus;
+import twitter4j.ResponseList;
 import twitter4j.Twitter;
 import twitter4j.TwitterException;
 import twitter4j.User;
 import util.DateUtil;
 import util.LogUtil;
+import util.UnfollowUtil;
 import zappa_gg.Messages;
 import zappa_gg.ZappaBot;
 
@@ -31,21 +34,25 @@ public class FriendService {
   private static final Logger logger = Logger.getLogger(FriendService.class.getName());
 
   private static final String GAE_FRIEND_LIST = "/gae/datastore/TwitterFriend";
+  private static final String GAE_UNFOLLOW_LIST = "/gae/datastore/Unfollow";
   private static final String API_FOLLOWERS_LIST = "/followers/list";
   private static final String API_FRIENDS_LIST = "/friends/list";
+  private static final String API_USERS_LIST = "/users/lookup";
   private static final String LIMIT_STATUS_FOLLOWERS = "followers";
   private static final String LIMIT_STATUS_FRIENDS = "friends";
+  private static final String LIMIT_STATUS_USERS = "users";
   private static final String FOLLOW_LIST_NAME = "follow";
-  private static final int API_PAGE_SIZE = 100;
-  private static final int API_MAX_COUNT = 15;
-  private static final int GAE_PAGE_SIZE = 1000;
+  private static final int API_LIST_PAGE_SIZE = 100;
+  private static final int API_LOOKUP_PAGE_SIZE = 100;
+  private static final int API_LIST_MAX_COUNT = 7;
+  private static final int GAE_PAGE_SIZE = 500;
   private static final int REMOVE_DAYS = 30;
 
   // GAE の1日あたりの制限 エンティティ 読み込み数 50,000 書き込み数 20,000 削除数 20,000
-  // 1回の実行で、Twitter: 100 PageSize * 15 Count = 1,500 ( > DataStore: 1000 )
-  // Entity 書き込み 4時間毎に実行した場合、1日で 1,500 * 6 = 9,000
+  // 1回の実行で、Twitter: 100 PageSize * 8 Count = 800 ( > DataStore: 500 )
+  // Entity 書き込み 2時間毎に実行した場合、1日で 800 * 12 = 9,600
   // 1日あたりの制限の半分以下くらいに収まる。
-  // よって4時間毎に実行することとする。
+  // よって2時間毎に実行することとする。
 
   @FunctionalInterface
   private interface Function<T, R> {
@@ -66,7 +73,7 @@ public class FriendService {
       throws TwitterException {
     // API制限取得
     final RateLimitStatus rateLimitStatus = tw.getRateLimitStatus(statusName).get(apiName);
-    final int limit = Math.min(rateLimitStatus.getRemaining(), API_MAX_COUNT);
+    final int limit = Math.min(rateLimitStatus.getRemaining(), API_LIST_MAX_COUNT);
     // カーソル初期化
     long cursor = PagableResponseList.START;
     NextCursorDao nextCursorDao = new NextCursorDao();
@@ -102,7 +109,8 @@ public class FriendService {
     final TwitterFriendDao dao = new TwitterFriendDao();
     // フォロワー
     long cur = loadList(tw, LIMIT_STATUS_FOLLOWERS, API_FOLLOWERS_LIST, (Long cursor) -> {
-      PagableResponseList<User> users = tw.getFollowersList(ZappaBot.SCREEN_NAME, cursor, API_PAGE_SIZE, true, false);
+      PagableResponseList<User> users = tw.getFollowersList(ZappaBot.SCREEN_NAME, cursor, API_LIST_PAGE_SIZE, true,
+          false);
       users.stream().forEach(user -> dao.saveLastFollowedByDate(user.getId(), date));
       return users.getNextCursor();
     });
@@ -121,7 +129,8 @@ public class FriendService {
     final TwitterFriendDao dao = new TwitterFriendDao();
     // フォロイー
     long cur = loadList(tw, LIMIT_STATUS_FRIENDS, API_FRIENDS_LIST, (Long cursor) -> {
-      PagableResponseList<User> users = tw.getFriendsList(ZappaBot.SCREEN_NAME, cursor, API_PAGE_SIZE, true, false);
+      PagableResponseList<User> users = tw.getFriendsList(ZappaBot.SCREEN_NAME, cursor, API_LIST_PAGE_SIZE, true,
+          false);
       users.stream().forEach(user -> dao.saveLastFollowingDate(user.getId(), date));
       return users.getNextCursor();
     });
@@ -136,8 +145,9 @@ public class FriendService {
    * @throws TwitterException
    */
   public Map<String, RateLimitStatus> loadRateLimitStatus(Twitter tw) throws TwitterException {
-    return tw.getRateLimitStatus(LIMIT_STATUS_FOLLOWERS, LIMIT_STATUS_FRIENDS).entrySet().stream()
-        .filter(e -> e.getKey().equals(API_FOLLOWERS_LIST) || e.getKey().equals(API_FRIENDS_LIST))
+    return tw.getRateLimitStatus(LIMIT_STATUS_FOLLOWERS, LIMIT_STATUS_FRIENDS, LIMIT_STATUS_USERS).entrySet().stream()
+        .filter(e -> e.getKey().equals(API_FOLLOWERS_LIST) || e.getKey().equals(API_FRIENDS_LIST)
+            || e.getKey().equals(API_USERS_LIST))
         .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
   }
 
@@ -166,7 +176,7 @@ public class FriendService {
     }
     // リムーブ対象外リスト
     Set<Long> followList = tw
-        .getUserListMembers(ZappaBot.SCREEN_NAME, FOLLOW_LIST_NAME, API_PAGE_SIZE, PagableResponseList.START, true)
+        .getUserListMembers(ZappaBot.SCREEN_NAME, FOLLOW_LIST_NAME, API_LIST_PAGE_SIZE, PagableResponseList.START, true)
         .stream().map(User::getId).collect(Collectors.toSet());
     // Twitterステータス更新用
     TwitterFriendDao twitterFriendDao = new TwitterFriendDao();
@@ -176,12 +186,16 @@ public class FriendService {
     String newCursor = null;
     try {
       while (iter.hasNext()) {
+        hasNext = true;
         TwitterFriend user = iter.next();
         try {
           if (user.getLastFollowedByDate() != null && user.getLastFollowingDate() == null) {
             // フォローされている & フォローしてない場合、フォロー
-            tw.createFriendship(user.getId());
-            logFollowCount++;
+            // ただし、アンフォロー判定を考慮する
+            if (!user.isUnfollow()) {
+              tw.createFriendship(user.getId());
+              logFollowCount++;
+            }
           } else if ((user.getLastFollowedByDate() == null && user.getLastFollowingDate() != null)
               || (user.getLastFollowedByDate() != null
                   && DateUtil.addDays(user.getLastFollowedByDate(), REMOVE_DAYS).compareTo(date) < 0)) {
@@ -196,7 +210,17 @@ public class FriendService {
           } else if (user.getLastFollowingDate() != null
               && DateUtil.addDays(user.getLastFollowingDate(), REMOVE_DAYS).compareTo(date) < 0) {
             // 一定期間フォローしていない場合、DBから削除
-            twitterFriendDao.delete(user.getId());
+            // ただし、アンフォロー判定を考慮する
+            if (!user.isUnfollow()) {
+              twitterFriendDao.delete(user.getId());
+            }
+          }
+          // フォロー判定
+          if (user.getLastFollowingDate() != null && user.isUnfollow() && !followList.contains(user.getId())) {
+            // フォローしている & アンフォロー判定 & フォローリストにない場合、リムーブを実行。
+            tw.destroyFriendship(user.getId());
+            twitterFriendDao.unfollow(user);
+            logRemoveCount++;
           }
         } catch (TwitterException e) {
           // ユーザーが見つからない場合、DBから削除
@@ -209,7 +233,6 @@ public class FriendService {
             throw e;
           }
         }
-        hasNext = true;
       }
     } finally {
       // カーソル位置を保存
@@ -223,6 +246,88 @@ public class FriendService {
       }
     }
     logger.info(String.format("follow:%d remove:%d", logFollowCount, logRemoveCount));
+    return newCursor == null || newCursor.isEmpty();
+  }
+
+  /**
+   * ツイッターからユーザー情報を取得し、フォロースべきか判定結果をDBに書き込む。
+   *
+   * @param tw
+   * @return true: 完了. false: 残あり.
+   * @throws TwitterException
+   */
+  public boolean updateUnfollow(Twitter tw) throws TwitterException {
+    // ログ用
+    int logUnfollowCount = 0;
+    // Function
+    TwitterFriendDao twitterFriendDao = new TwitterFriendDao();
+    UnfollowUtil unfollowUtil = new UnfollowUtil();
+    Function<Map<Long, TwitterFriend>, Integer> detectUnfollow = friends -> {
+      int cnt = 0;
+      try {
+        ResponseList<User> users = tw.lookupUsers(friends.keySet().stream().mapToLong(Long::longValue).toArray());
+        for (User user : users) {
+          if (unfollowUtil.detectUnfollow(user)) {
+            twitterFriendDao.updateUnfollow(friends.get(user.getId()), true);
+            cnt++;
+          }
+        }
+      } catch (TwitterException e) {
+        // statusCode=404, message=No user matches for specified terms., code=17
+        if (e.getErrorCode() == 17) {
+          // ignore
+        } else {
+          LogUtil.sendDirectMessage(tw, String.format("%s: Fail unfollow, statusCode:%d code:%d message:%s",
+              Messages.ERROR_MESSAGE, e.getStatusCode(), e.getErrorCode(), e.getMessage()));
+          throw e;
+        }
+      }
+      return cnt;
+    };
+    // カーソル初期化
+    Query<TwitterFriend> query = ofy().load().type(TwitterFriend.class).limit(GAE_PAGE_SIZE);
+    NextCursorDao nextCursorDao = new NextCursorDao();
+    NextCursor nextCursor = nextCursorDao.loadById(GAE_UNFOLLOW_LIST);
+    String cursor = null;
+    if (nextCursor != null) {
+      cursor = nextCursor.getNextCursor();
+      if (cursor != null) {
+        query = query.startAt(Cursor.fromWebSafeString(nextCursor.getNextCursor()));
+      }
+    }
+    // DBから取得。
+    QueryResultIterator<TwitterFriend> iter = query.iterator();
+    boolean hasNext = false; // 次ページがあるか
+    String newCursor = null;
+    try {
+      Map<Long, TwitterFriend> friends = new HashMap<>();
+      while (iter.hasNext()) {
+        hasNext = true;
+        TwitterFriend friend = iter.next();
+        if (friend.isUnfollow()) {
+          continue;
+        }
+        friends.put(friend.getId(), friend);
+        if (friends.size() == API_LOOKUP_PAGE_SIZE) {
+          logUnfollowCount += detectUnfollow.apply(friends);
+          friends.clear();
+        }
+      }
+      if (0 < friends.size() && friends.size() < API_LOOKUP_PAGE_SIZE) {
+        logUnfollowCount += detectUnfollow.apply(friends);
+      }
+    } finally {
+      // カーソル位置を保存
+      if (hasNext) {
+        newCursor = iter.getCursor().toWebSafeString();
+      }
+      if (nextCursor == null) {
+        nextCursorDao.insertNextCursor(GAE_UNFOLLOW_LIST, newCursor);
+      } else {
+        nextCursorDao.updateNextCursor(nextCursor, newCursor);
+      }
+    }
+    logger.info(String.format("unfollow:%d", logUnfollowCount));
     return newCursor == null || newCursor.isEmpty();
   }
